refactor(agent): extract contribution pr service
