+++ src/orchestration/agent.ts
@@ -1,4 +1,4 @@
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
+import { existsSync, mkdirSync, writeFileSync } from 'fs';
 import { homedir } from 'os';
 import { join } from 'path';
 import { Octokit } from '@octokit/rest';
@@ -25,6 +25,7 @@ import {
 } from '../infra/index.js';
 import {
   contentService,
+  contributionPrService,
   gitService,
   githubService,
   inboxService,
@@ -55,11 +56,6 @@ interface TargetRepoContext {
   owner: string;
   repo: string;
   defaultBranch: string;
-}
-
-interface DraftPullRequest {
-  title: string;
-  body: string;
 }
 
 interface ContributionPullRequestResult {
@@ -756,6 +752,7 @@ export class AgentOrchestrator {
     }
 
     this.octokit = new Octokit({ auth: config.github.pat });
+    contributionPrService.initialize(this.octokit);
     if (!validateLlm) {
       return;
     }
@@ -1236,21 +1233,12 @@ export class AgentOrchestrator {
     }
 
     try {
-      const upstreamRepo = await this.getUpstreamRepositoryContext(input.issue);
-      const forkRepo = await this.ensureForkRepository(upstreamRepo);
-      const branchName = this.buildPublishBranchName(input.issue);
-      const draftPullRequest = this.buildDraftPullRequest(input.prDraft);
-      const commitMessage = this.buildContributionCommitMessage(input.issue);
-
-      await this.createCommitOnFork({
-        forkRepo,
-        branchName,
+      const contributionPullRequest = await contributionPrService.submitDraftPullRequest({
+        issue: input.issue,
+        prDraft: input.prDraft,
         workspacePath: input.workspace.workspacePath,
         changedFiles: input.changedFiles,
-        commitMessage,
       });
-
-      const contributionPullRequest = await this.createContributionPullRequest(upstreamRepo, forkRepo.owner, branchName, draftPullRequest);
 
       ui.card({
         label: 'OpenMeta Agent',
@@ -1258,7 +1246,7 @@ export class AgentOrchestrator {
         subtitle: 'The generated patch has been pushed to your fork and turned into a real upstream draft PR.',
         lines: [
           `Repository: ${input.issue.repoFullName}`,
-          `Branch: ${branchName}`,
+          `Branch: ${contributionPullRequest.branchName}`,
           `Changed Files: ${input.changedFiles.join(', ')}`,
           `Pull Request: ${contributionPullRequest.url}`,
         ],
@@ -1266,7 +1254,7 @@ export class AgentOrchestrator {
       });
 
       return {
-        branchName,
+        branchName: contributionPullRequest.branchName,
         url: contributionPullRequest.url,
         number: contributionPullRequest.number,
         changedFiles: input.changedFiles,
@@ -1701,23 +1689,6 @@ export class AgentOrchestrator {
     return data;
   }
 
-  private buildPublishBranchName(issue: RankedIssue): string {
-    const slug = issue.title
-      .toLowerCase()
-      .replace(/[^a-z0-9]+/g, '-')
-      .replace(/^-+|-+$/g, '')
-      .slice(0, 32);
-
-    return `openmeta/agent-${issue.number}-${slug || 'issue'}-${Date.now()}`;
-  }
-
-  private buildDraftPullRequest(prDraft: PullRequestDraft): DraftPullRequest {
-    return {
-      title: prDraft.title,
-      body: contentService.formatPullRequestDraftBody(prDraft),
-    };
-  }
-
   private showStructuredReviewNotice(input: {
     title: string;
     subtitle: string;
@@ -1729,223 +1700,6 @@ export class AgentOrchestrator {
       subtitle: input.subtitle,
       lines: input.lines,
       tone: 'warning',
-    });
-  }
-
-  private async getUpstreamRepositoryContext(issue: RankedIssue): Promise<TargetRepoContext> {
-    const [owner, repo] = issue.repoFullName.split('/');
-    if (!owner || !repo) {
-      throw new Error(`Invalid issue repository reference: ${issue.repoFullName}`);
-    }
-
-    const repoInfo = await this.getGitHubRepositoryInfo(owner, repo);
-    return {
-      path: '',
-      owner,
-      repo,
-      defaultBranch: repoInfo.default_branch || 'main',
-    };
-  }
-
-  private async ensureForkRepository(upstreamRepo: TargetRepoContext): Promise<TargetRepoContext> {
-    if (!this.octokit) {
-      throw new Error('GitHub service not initialized');
-    }
-
-    const forkOwner = githubService.getUsername();
-
-    try {
-      const { data } = await this.octokit.rest.repos.get({
-        owner: forkOwner,
-        repo: upstreamRepo.repo,
-      });
-
-      if (!data.fork || data.parent?.full_name !== `${upstreamRepo.owner}/${upstreamRepo.repo}`) {
-        throw new Error(`Repository ${forkOwner}/${upstreamRepo.repo} exists but is not a fork of ${upstreamRepo.owner}/${upstreamRepo.repo}.`);
-      }
-
-      await this.syncForkWithUpstream(forkOwner, upstreamRepo.repo, data.default_branch || upstreamRepo.defaultBranch);
-      return {
-        path: '',
-        owner: forkOwner,
-        repo: upstreamRepo.repo,
-        defaultBranch: data.default_branch || upstreamRepo.defaultBranch,
-      };
-    } catch (error) {
-      const err = error as { status?: number };
-      if (err.status && err.status !== 404) {
-        throw error;
-      }
-    }
-
-    logger.info(`Creating fork for ${upstreamRepo.owner}/${upstreamRepo.repo}`);
-    await this.octokit.rest.repos.createFork({
-      owner: upstreamRepo.owner,
-      repo: upstreamRepo.repo,
-    });
-
-    const fork = await this.waitForFork(forkOwner, upstreamRepo.repo, `${upstreamRepo.owner}/${upstreamRepo.repo}`);
-    await this.syncForkWithUpstream(forkOwner, upstreamRepo.repo, fork.default_branch || upstreamRepo.defaultBranch);
-
-    return {
-      path: '',
-      owner: forkOwner,
-      repo: upstreamRepo.repo,
-      defaultBranch: fork.default_branch || upstreamRepo.defaultBranch,
-    };
-  }
-
-  private async waitForFork(owner: string, repo: string, expectedParent: string) {
-    if (!this.octokit) {
-      throw new Error('GitHub service not initialized');
-    }
-
-    for (let attempt = 0; attempt < 8; attempt += 1) {
-      try {
-        const { data } = await this.octokit.rest.repos.get({ owner, repo });
-        if (data.fork && data.parent?.full_name === expectedParent) {
-          return data;
-        }
-      } catch {
-        // Continue polling until the fork is visible.
-      }
-
-      await this.delay(1500);
-    }
-
-    throw new Error(`Fork ${owner}/${repo} was not ready in time.`);
-  }
-
-  private async syncForkWithUpstream(owner: string, repo: string, branch: string): Promise<void> {
-    if (!this.octokit) {
-      throw new Error('GitHub service not initialized');
-    }
-
-    try {
-      await this.octokit.rest.repos.mergeUpstream({
-        owner,
-        repo,
-        branch,
-      });
-    } catch (error) {
-      logger.debug(`Unable to sync fork ${owner}/${repo} with upstream before opening a PR`, error);
-    }
-  }
-
-  private async createCommitOnFork(input: {
-    forkRepo: TargetRepoContext;
-    branchName: string;
-    workspacePath: string;
-    changedFiles: string[];
-    commitMessage: string;
-  }): Promise<void> {
-    if (!this.octokit) {
-      throw new Error('GitHub service not initialized');
-    }
-
-    const branch = await this.octokit.rest.repos.getBranch({
-      owner: input.forkRepo.owner,
-      repo: input.forkRepo.repo,
-      branch: input.forkRepo.defaultBranch,
-    });
-
-    const baseCommitSha = branch.data.commit.sha;
-    const baseTreeSha = branch.data.commit.commit.tree.sha;
-
-    const tree = input.changedFiles.map((filePath) => ({
-      path: filePath,
-      mode: '100644' as const,
-      type: 'blob' as const,
-      content: readFileSync(join(input.workspacePath, filePath), 'utf-8'),
-    }));
-
-    const createdTree = await this.octokit.rest.git.createTree({
-      owner: input.forkRepo.owner,
-      repo: input.forkRepo.repo,
-      base_tree: baseTreeSha,
-      tree,
-    });
-
-    const createdCommit = await this.octokit.rest.git.createCommit({
-      owner: input.forkRepo.owner,
-      repo: input.forkRepo.repo,
-      message: input.commitMessage,
-      tree: createdTree.data.sha,
-      parents: [baseCommitSha],
-    });
-
-    try {
-      await this.octokit.rest.git.createRef({
-        owner: input.forkRepo.owner,
-        repo: input.forkRepo.repo,
-        ref: `refs/heads/${input.branchName}`,
-        sha: createdCommit.data.sha,
-      });
-    } catch (error) {
-      const err = error as { status?: number };
-      if (err.status !== 422) {
-        throw error;
-      }
-
-      await this.octokit.rest.git.updateRef({
-        owner: input.forkRepo.owner,
-        repo: input.forkRepo.repo,
-        ref: `heads/${input.branchName}`,
-        sha: createdCommit.data.sha,
-        force: true,
-      });
-    }
-  }
-
-  private async createContributionPullRequest(
-    upstreamRepo: TargetRepoContext,
-    forkOwner: string,
-    branchName: string,
-    draftPullRequest: DraftPullRequest,
-  ): Promise<{ url: string; number: number }> {
-    if (!this.octokit) {
-      throw new Error('GitHub service not initialized');
-    }
-
-    const existing = await this.octokit.rest.pulls.list({
-      owner: upstreamRepo.owner,
-      repo: upstreamRepo.repo,
-      head: `${forkOwner}:${branchName}`,
-      base: upstreamRepo.defaultBranch,
-      state: 'open',
-    });
-
-    const [existingPullRequest] = existing.data;
-    if (existingPullRequest) {
-      return {
-        url: existingPullRequest.html_url,
-        number: existingPullRequest.number,
-      };
-    }
-
-    const { data } = await this.octokit.rest.pulls.create({
-      owner: upstreamRepo.owner,
-      repo: upstreamRepo.repo,
-      title: draftPullRequest.title,
-      body: draftPullRequest.body,
-      head: `${forkOwner}:${branchName}`,
-      base: upstreamRepo.defaultBranch,
-      draft: true,
-    });
-
-    return {
-      url: data.html_url,
-      number: data.number,
-    };
-  }
-
-  private buildContributionCommitMessage(issue: RankedIssue): string {
-    return `feat: address ${issue.repoFullName}#${issue.number} ${issue.title}`.slice(0, 120);
-  }
-
-  private delay(ms: number): Promise<void> {
-    return new Promise((resolve) => {
-      setTimeout(resolve, ms);
     });
   }
